fix(query): avoid update-time sleep dating
